@@ -1,5 +1,6 @@
 import React from "react";
-import { CartItem } from "../types";
+import type { CartItem } from "../types";
+import { FaBagShopping } from "react-icons/fa6";
 
 interface CartProps {
   cartItems: CartItem[];
@@ -19,44 +20,19 @@ const Cart: React.FC<CartProps> = ({
 
   return (
     <div className="max-w-4xl mx-auto p-6">
-      <div className="bg-white rounded-2xl shadow-soft border border-gray-100 overflow-hidden">
+      <div className="bg-white rounded-2xl shadow-lg border border-gray-100 overflow-hidden">
         {/* Header */}
-        <div className="bg-gradient-to-r from-primary-600 to-primary-700 px-6 py-4">
+        <div className="bg-gradient-to-r from-blue-600 to-blue-700 px-6 py-4">
           <h2 className="text-2xl font-bold text-white flex items-center gap-3">
-            <svg
-              className="w-6 h-6"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M3 3h2l.4 2M7 13h10l4-8H5.4m0 0L7 13m0 0l-2.5 5M7 13l2.5 5m6-5v6a2 2 0 01-2 2H9a2 2 0 01-2-2v-6m8 0V9a2 2 0 00-2-2H9a2 2 0 00-2 2v4.01"
-              />
-            </svg>
+            <FaBagShopping className="w-6 h-6" />
             Shopping Cart
           </h2>
         </div>
 
-        {/* Cart Items */}
         <div className="p-6">
           {cartItems.length === 0 ? (
             <div className="text-center py-12">
-              <svg
-                className="w-16 h-16 text-gray-300 mx-auto mb-4"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M3 3h2l.4 2M7 13h10l4-8H5.4m0 0L7 13m0 0l-2.5 5M7 13l2.5 5m6-5v6a2 2 0 01-2 2H9a2 2 0 01-2-2v-6m8 0V9a2 2 0 00-2-2H9a2 2 0 00-2 2v4.01"
-                />
-              </svg>
+              <FaBagShopping className="w-16 h-16 text-gray-300 mx-auto mb-4" />
               <p className="text-gray-500 text-lg font-medium">
                 Your cart is empty
               </p>
@@ -71,7 +47,6 @@ const Cart: React.FC<CartProps> = ({
                   key={item.id}
                   className="flex items-center gap-4 p-4 bg-gray-50 rounded-xl border border-gray-100"
                 >
-                  {/* Product Image */}
                   <div className="flex-shrink-0">
                     <img
                       src={item.product?.imageUrl}
@@ -80,17 +55,15 @@ const Cart: React.FC<CartProps> = ({
                     />
                   </div>
 
-                  {/* Product Details */}
                   <div className="flex-1 min-w-0">
                     <h3 className="text-lg font-semibold text-gray-900 truncate">
                       {item.product?.name}
                     </h3>
-                    <p className="text-primary-600 font-semibold">
+                    <p className="text-blue-600 font-semibold">
                       ${item.product?.price.toFixed(2)}
                     </p>
                   </div>
 
-                  {/* Quantity Controls */}
                   <div className="flex items-center gap-3">
                     <div className="flex items-center bg-white border border-gray-200 rounded-lg">
                       <button
@@ -162,7 +135,6 @@ const Cart: React.FC<CartProps> = ({
                     </button>
                   </div>
 
-                  {/* Item Total */}
                   <div className="text-right">
                     <p className="text-lg font-bold text-gray-900">
                       ${((item.product?.price || 0) * item.quantity).toFixed(2)}
@@ -174,13 +146,12 @@ const Cart: React.FC<CartProps> = ({
           )}
         </div>
 
-        {/* Cart Total */}
         {cartItems.length > 0 && (
           <div className="border-t border-gray-200 bg-gray-50 px-6 py-4">
             <div className="flex items-center justify-between">
               <h3 className="text-xl font-semibold text-gray-900">Total</h3>
               <div className="text-right">
-                <p className="text-3xl font-bold text-primary-600">
+                <p className="text-3xl font-bold text-blue-600">
                   ${total.toFixed(2)}
                 </p>
                 <p className="text-sm text-gray-500 mt-1">
