@@ -1,5 +1,6 @@
 import React from "react";
-import { Product } from "../types";
+import type { Product } from "../types";
+import { FaBagShopping } from "react-icons/fa6";
 
 interface ProductCardProps {
   product: Product;
@@ -8,8 +9,7 @@ interface ProductCardProps {
 
 const ProductCard: React.FC<ProductCardProps> = ({ product, onAddToCart }) => {
   return (
-    <div className="group relative bg-white rounded-2xl shadow-soft hover:shadow-medium transition-all duration-300 overflow-hidden border border-gray-100">
-      {/* Product Image */}
+    <div className="group relative bg-white rounded-2xl shadow-lg hover:shadow-xl transition-all duration-300 overflow-hidden border border-gray-100">
       <div className="aspect-square overflow-hidden bg-gray-100">
         <img
           src={product.imageUrl}
@@ -18,7 +18,6 @@ const ProductCard: React.FC<ProductCardProps> = ({ product, onAddToCart }) => {
         />
       </div>
 
-      {/* Product Info */}
       <div className="p-6">
         <h3 className="text-lg font-semibold text-gray-900 mb-2 line-clamp-2">
           {product.name}
@@ -27,37 +26,23 @@ const ProductCard: React.FC<ProductCardProps> = ({ product, onAddToCart }) => {
           {product.description}
         </p>
 
-        {/* Price and Action */}
         <div className="flex items-center justify-between">
           <div className="flex items-baseline gap-1">
-            <span className="text-2xl font-bold text-primary-600">
+            <span className="text-2xl font-bold text-blue-600">
               ${product.price.toFixed(2)}
             </span>
           </div>
 
           <button
             onClick={() => onAddToCart(product.id)}
-            className="btn-primary flex items-center gap-2 px-4 py-2 text-sm"
+            className="btn-primary flex items-center gap-1 px-2 py-1 text-xs"
           >
-            <svg
-              className="w-4 h-4"
-              fill="none"
-              stroke="currentColor"
-              viewBox="0 0 24 24"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M3 3h2l.4 2M7 13h10l4-8H5.4m0 0L7 13m0 0l-2.5 5M7 13l2.5 5m6-5v6a2 2 0 01-2 2H9a2 2 0 01-2-2v-6m8 0V9a2 2 0 00-2-2H9a2 2 0 00-2 2v4.01"
-              />
-            </svg>
+            <FaBagShopping className="w-4 h-4" />
             Add to Cart
           </button>
         </div>
       </div>
 
-      {/* Hover Effect Overlay */}
       <div className="absolute inset-0 bg-gradient-to-t from-black/20 to-transparent opacity-0 group-hover:opacity-100 transition-opacity duration-300 pointer-events-none" />
     </div>
   );
